+++ pillar/documents.go
@@ -1,0 +1,5 @@
+package pillar
+
+type storeResponse struct {
+	Total int `json:"total_sold"`
+}